move maybe_detect_* helpers to utils/async_probe
