+++ python/sglang/srt/utils/async_probe.py
@@ -1,0 +1,48 @@
+"""Async invariant probes — fire torch._assert_async without CPU sync.
+
+All probes are gated on SGLANG_ENABLE_ASYNC_ASSERT (default off in prod).
+When the gate is on, a violation surfaces as an assertion at the next CUDA
+sync point instead of as a silent NaN cascade or illegal-address crash.
+"""
+
+import torch
+
+from sglang.srt.environ import envs
+
+
+def maybe_detect_nan(tensor: torch.Tensor, msg: str = ""):
+    """Async NaN check — no GPU-CPU sync, error surfaces at next sync point."""
+    if not envs.SGLANG_ENABLE_ASYNC_ASSERT.get():
+        return
+    torch._assert_async(~torch.any(torch.isnan(tensor)), f"NaN detected! {msg}")
+
+
+def maybe_detect_inf(tensor: torch.Tensor, msg: str = ""):
+    """Async Inf check — fp16 overflow surfaces as Inf before NaN."""
+    if not envs.SGLANG_ENABLE_ASYNC_ASSERT.get():
+        return
+    torch._assert_async(~torch.any(torch.isinf(tensor)), f"Inf detected! {msg}")
+
+
+def maybe_detect_oob(indices: torch.Tensor, low: int, high: int, msg: str):
+    """Async OOB check — no GPU-CPU sync, error surfaces at next sync point."""
+    if not envs.SGLANG_ENABLE_ASYNC_ASSERT.get():
+        return
+    if indices.numel() == 0:
+        return
+    torch._assert_async(
+        (indices.min() >= low) & (indices.max() < high),
+        f"OOB indices not in [{low}, {high}): {msg}",
+    )
+
+
+def maybe_detect_page_aligned(indices: torch.Tensor, page_size: int, msg: str):
+    """Async page-alignment check on slot ids."""
+    if not envs.SGLANG_ENABLE_ASYNC_ASSERT.get():
+        return
+    if indices.numel() == 0 or page_size <= 1:
+        return
+    torch._assert_async(
+        (indices % page_size == 0).all(),
+        f"page-misaligned indices (page_size={page_size}): {msg}",
+    )
